# openapi.yaml
openapi: 3.0.3
info:
  title: Fly-Fleet API
  description: Private Jet Charter Brokerage Platform API
  version: 1.0.0
  contact:
    name: Fly-Fleet Support
    email: contact@fly-fleet.com
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/admin/email-activities:
    get:
      summary: List admin
      description: Administrative operations requiring authentication
      tags:
        - admin
        - admin
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
          description: Pagination offset
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
  /api/admin/email-stats:
    get:
      summary: List admin
      description: Administrative operations requiring authentication
      tags:
        - admin
        - admin
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
  /api/admin/quotes:
    get:
      summary: List admin
      description: Administrative operations requiring authentication
      tags:
        - admin
        - admin
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - AdminAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
          description: Pagination offset
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
    post:
      summary: Create admin
      description: Administrative operations requiring authentication
      tags:
        - admin
        - admin
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - AdminAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
          description: Pagination offset
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
...
  /api/quotes/{id}/status:
    get:
      summary: Get quotes
      description: Manage quote requests for private jet charter services
      tags:
        - quotes
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Id parameter
    patch:
      summary: Partially update quotes
      description: Manage quote requests for private jet charter services
      tags:
        - quotes
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteStatusUpdateSchema'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Id parameter
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin authentication using Bearer token with x-admin-email header
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    path:
                      type: array
                      items:
                        oneOf:
                          - type: string
                          - type: integer
                    message:
                      type: string
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication required
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error